+++ ontimize/components/contextmenu/context-menu-group/o-context-menu-group.component.ts
@@ -4,7 +4,8 @@ import { OContextMenuItemComponent, DEFAULT_INPUTS_O_CONTEXT_MENU_ITEM } from '.
 
 export const DEFAULT_CONTEXT_MENU_GROUP_INPUTS = [
   ...DEFAULT_INPUTS_O_CONTEXT_MENU_ITEM,
-  'children'];
+  'children'
+];
 
 @Component({
   moduleId: module.id,
@@ -14,16 +15,12 @@ export const DEFAULT_CONTEXT_MENU_GROUP_INPUTS = [
   providers: [{ provide: OComponentMenuItems, useExisting: forwardRef(() => OContextMenuGroupComponent) }]
 })
 
-export class OContextMenuGroupComponent extends OContextMenuItemComponent  implements OnInit {
+export class OContextMenuGroupComponent extends OContextMenuItemComponent implements OnInit {
 
   public type = OComponentMenuItems.TYPE_GROUP_MENU;
   public children = [];
 
   @ContentChildren(OComponentMenuItems) public oContextMenuItems: QueryList<OComponentMenuItems>;
-
-  ngOnInit() {
-    super.ngOnInit();
-  }
 
   ngAfterContentInit() {
     this.children = this.oContextMenuItems.toArray().slice(1, this.oContextMenuItems.toArray().length);